% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dverse-package.R
\docType{package}
\name{dverse-package}
\alias{dverse}
\alias{dverse-package}
\title{dverse: Document a Universe of Packages}
\description{
Creates a data frame of the metadata associated to the documentation of a collection of R packages. If you maintain universe meta-package, it helps you to easily create a universal reference for its website.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maurolepore/dverse}
  \item \url{https://maurolepore.github.io/dverse/}
  \item Report bugs at \url{https://github.com/maurolepore/dverse/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com} [contractor]

}
\keyword{internal}
