% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_online.R
\name{is_online}
\alias{is_online}
\title{Is each URL (or href) online?}
\usage{
is_online(x)
}
\arguments{
\item{x}{String.}
}
\value{
A logical vector.
}
\description{
Is each URL (or href) online?
}
\examples{
url <- c(
  "https://example.com",
  "<a href=https://example.com>example</a>",
  "<a href=https://example.com/bad>bad</a>",
  "bad"
)

is_online(url)
}
