% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_universe.R
\name{document_universe}
\alias{document_universe}
\title{Create a data frame with documentation metadata of one or more packages}
\usage{
document_universe(x, url_template = NULL)
}
\arguments{
\item{x}{A character vector giving concepts or package names to match.}

\item{url_template}{Character. A template to generate links to documentation,
using the syntax of \code{glue::glue()} to indicate where to insert the values
from the columns \code{package} and \code{topic}.

The template may have lengh 1 or 2:
\itemize{
\item If the template is a vector has length 1, the template should point to the
manual (a.k.a. "reference" or "index"), i.e. the online version of the help
files that you can otherwise access locally with \verb{?}, e.g.:
}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{"https://\{package\}.tidyverse.org/reference/\{topic\}.html"
}\if{html}{\out{</div>}}

In this case, the template for vignettes is automatically constructed by
replacing \verb{/reference/} with \verb{/articles/}, e.g.:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{"https://\{package\}.tidyverse.org/articles/\{topic\}.html"
}\if{html}{\out{</div>}}

If this is invalid, then you'll need to provide a vector of length 2.
\itemize{
\item If the template is a vector has length 2, the first element should point to the manual
(see above), and the second element should point to vignettes, e.g.:
}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{c(
 "https://\{package\}.tidyverse.org/reference/\{topic\}.html",
 "https://\{package\}.tidyverse.org/articles/\{topic\}.html"
)
}\if{html}{\out{</div>}}}
}
\value{
A data frame.
}
\description{
Create a data.frame with documentation metadata of one or more packages. If
you develop a universe, such as the tidyverse or tidymodels, dverse helps you
to easily create a universe-wide reference for the pkgdown website of your
meta-package.
}
\examples{
library(glue)
library(tibble)

universe <- c("glue", "tibble")

document_universe(universe)

# Adding links the online manual and vignettes
template <- "https://{package}.tidyverse.org/reference/{topic}.html"
document_universe(universe, template)

# Adding an explicit template for vignettes
template <- c(
 "https://{package}.tidyverse.org/reference/{topic}.html",
 "https://{package}.tidyverse.org/articles/{topic}.html"
)
document_universe(universe, template)

# Works beyond GitHub Pages, e.g. on r-universe
template <- c(
  "https://tidyverse.r-universe.dev/{package}/doc/manual.html#{topic}",
  "https://tidyverse.r-universe.dev/articles/{package}/{topic}.html"
)
document_universe(universe, template)
}
