% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dverse-package.R
\docType{package}
\name{dverse-package}
\alias{dverse}
\alias{dverse-package}
\title{dverse: Document a Universe of Packages}
\description{
Creates a data frame containing the metadata associated with the documentation of a collection of R packages. It allows for linking topic names to their corresponding documentation online. If you maintain a universe meta-package, it helps create a comprehensive reference for its website.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maurolepore/dverse}
  \item \url{https://maurolepore.github.io/dverse/}
  \item Report bugs at \url{https://github.com/maurolepore/dverse/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com} (\href{https://orcid.org/0000-0002-1986-7988}{ORCID}) [copyright holder]

}
\keyword{internal}
